debug

+++ modules/menus.py
@@ -215,6 +215,14 @@ def debug_submenu():
         case 18:
             net.install_networking_packages()
             print('Das System muss neugestartet werden, fortfahren?')
-            os.system('sudo reboot')
+            restart_needed = ''
+            # Only 'y' and 'n' allowed
+            while restart_needed not in ['Y','N']:
+                restart_needed = input('(y/n): ').upper()
+            restart_needed = True if restart_needed == 'Y' else False
+            if restart_needed is True:
+                os.system('sudo reboot')
+            else:
+                debug_submenu()
         case _:
             debug_submenu()
